% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rodano_api_communication.R
\name{add_user}
\alias{add_user}
\title{Create new user}
\usage{
add_user(
  name,
  email,
  profile,
  parentPk,
  urlBase,
  auth,
  activate = FALSE,
  pwd = NULL
)
}
\arguments{
\item{name}{Character. Name of the user.}

\item{email}{Character. Email address of the user.}

\item{profile}{Character. Profile ID to assign to the user.}

\item{parentPk}{Integer. Primary key of the parent scope to attach the user to.}

\item{urlBase}{Character. The base URL to the platform's API.}

\item{auth}{An authentication object created by \code{\link{create_authentication}}.}

\item{activate}{Logical. Whether the user must be enabled immediately
(default: FALSE).}

\item{pwd}{Character or NULL. Password for the user. When NULL (default),
no password is set and user is not automatically enabled.}
}
\value{
Parsed response content from user creation request.
}
\description{
Creates a new user in the platform. Optionally activates the user
account if password is provided.
}
