% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rodano_api_communication.R
\name{get_connection_token}
\alias{get_connection_token}
\title{Retrieve authentication token}
\usage{
get_connection_token(urlBase, email = NULL, pwd = NULL)
}
\arguments{
\item{urlBase}{Character. The base URL to the platform's API.}

\item{email}{Character or NULL. The email used for login. If NULL,
prompts the user for input.}

\item{pwd}{Character or NULL. The password used for login. If NULL,
prompts the user for secure input.}
}
\value{
A character string containing the authentication token.
}
\description{
Retrieves an authentication token for a given platform using specific
credentials. Prompts for email and password if not provided.
}
