% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rodano_api_communication.R
\name{get_overdue_widget_reports}
\alias{get_overdue_widget_reports}
\title{Retrieve multiple overdue widget reports}
\usage{
get_overdue_widget_reports(urlBase, auth, widgetIds)
}
\arguments{
\item{urlBase}{Character. The base URL to the platform's API.}

\item{auth}{An authentication object created by \code{\link{create_authentication}}.}

\item{widgetIds}{Character vector. Overdue widget names to retrieve.}
}
\value{
A named list of data frames containing the overdue widget reports.
  List names are the lowercase widget IDs.
}
\description{
Retrieves multiple overdue widget reports from the server in a batch
operation using \code{\link{get_overdue_widget}}. Returns a named list
where each element is a data frame from one overdue widget.
}
\seealso{
\code{\link{get_overdue_widget}} for retrieving a single overdue widget report.
}
