% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rodano_api_communication.R
\name{get_reports}
\alias{get_reports}
\title{Retrieve multiple workflow reports}
\usage{
get_reports(urlBase, auth, reportIds, maxAttempts = 5, withHistory = FALSE)
}
\arguments{
\item{urlBase}{Character. The base URL to the platform's API.}

\item{auth}{An authentication object created by \code{\link{create_authentication}}.}

\item{reportIds}{Character vector. Workflow report names to retrieve.}

\item{maxAttempts}{Integer. Maximum number of retry attempts in case of failure (default: 5).}

\item{withHistory}{Logical. Should reports include historical data?}
}
\value{
A named list of data frames containing the workflow reports. List
  names are the lowercase report IDs.
}
\description{
Retrieves multiple workflow reports from the server in a batch operation using
\code{\link{get_report}}. Returns a named list where each element is a
data frame from one report.
}
\seealso{
\code{\link{get_report}} for retrieving a single workflow report.
}
