% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rodano_api_communication.R
\name{logout_user}
\alias{logout_user}
\title{Log out user session}
\usage{
logout_user(urlBase, auth)
}
\arguments{
\item{urlBase}{Character. The base URL to the platform's API.}

\item{auth}{An authentication object created by \code{\link{create_authentication}}.}
}
\value{
Called for its side effect. Throws an error if logout fails.
}
\description{
Deletes a user's session on the platform, invalidating the current
authentication token.
}
