% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_helper.R
\name{shift}
\alias{shift}
\title{Circular shift of vector elements}
\usage{
shift(x, i = 1)
}
\arguments{
\item{x}{A vector to shift.}

\item{i}{Integer number of positions to shift to the right (default: 1).
Negative values shift to the left. Values larger than \code{length(x)}
are taken modulo the vector length.}
}
\value{
A vector of the same length as \code{x} with elements rotated.
}
\description{
Rotates the elements of a vector to the right by \code{i} positions.
Elements that fall off the end wrap around to the beginning.
}
\examples{
shift(1:5) # c(5, 1, 2, 3, 4)
shift(1:5, 2) # c(4, 5, 1, 2, 3)
shift(1:5, -1) # c(2, 3, 4, 5, 1)

}
