% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rodano_api_communication.R
\name{get_connection_robot}
\alias{get_connection_robot}
\title{Retrieve robot credentials}
\usage{
get_connection_robot(urlBase, token, role = NULL, autoLogout = FALSE)
}
\arguments{
\item{urlBase}{Character. The base URL to the platform's API.}

\item{token}{Character. Authentication token of a user enabled to view robots.}

\item{role}{Character or NULL. ID of the role wanted for the robot. If NULL,
all robot roles are considered.}

\item{autoLogout}{Logical. If TRUE, logs out the user used for robot
retrieval (default: FALSE).}
}
\value{
A JSON robot object containing robot credentials and configuration.
}
\description{
Retrieves a robot account from a given platform. If multiple robots are
available or if the desired robot is ambiguous, the user will be prompted
to select from a list.
}
